Update: update version

+++ src/main/java/com/feliopolis/bookskeeper/controllers/BookController.java
@@ -1,9 +1,7 @@
 package com.feliopolis.bookskeeper.controllers;
 
 import com.feliopolis.bookskeeper.models.Book;
-import com.feliopolis.bookskeeper.models.User;
 import com.feliopolis.bookskeeper.repositories.BookRepository;
-import com.feliopolis.bookskeeper.repositories.UserRepository;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -29,31 +27,52 @@ public class BookController {
     @GetMapping
     @RequestMapping("{id}")
     public ResponseEntity<Book> get(@PathVariable Long id) {
-
         try {
-            return new ResponseEntity<Book>(bookRepository.getById(id), HttpStatus.OK);
-        } catch (ResponseStatusException exception) {
-            // TODO: is catch working?
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Book Not Found");
-        }
+            if (bookRepository.findById(id).isPresent())
+                return new ResponseEntity<Book>(bookRepository.getById(id), HttpStatus.OK);
+            else
+                return new ResponseEntity("Book not found", HttpStatus.NOT_FOUND);
 
+        } catch (ResponseStatusException exception) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Error getting book");
+        }
     }
+
 
     @PostMapping
     public Book create(@RequestBody final Book book) {
+
+        System.out.println("Book! " + book.getName());
+        System.out.println("Book! " + book.getDescription());
+        System.out.println("Book! " + book.getAuthor());
+        System.out.println("Book! " + book.getDate());
+
         return bookRepository.saveAndFlush(book);
+
     }
 
     @RequestMapping(value = "{id}", method = RequestMethod.DELETE)
-    public void delete(@PathVariable Long id) {
-        bookRepository.deleteById(id);
+    public ResponseEntity<Book> delete(@PathVariable Long id) {
+        if (bookRepository.findById(id).isPresent()) {
+            Book book = bookRepository.getById(id);
+            bookRepository.deleteById(id);
+            return new ResponseEntity<Book>(book, HttpStatus.OK);
+        } else
+            return new ResponseEntity("Book not found", HttpStatus.NOT_FOUND);
     }
 
     @RequestMapping(value = "{id}", method = RequestMethod.PUT)
-    public Book update(@PathVariable Long id, @RequestBody Book book) {
-        Book savedBook = bookRepository.getById(id);
-        BeanUtils.copyProperties(book, savedBook, "id");
-        return bookRepository.saveAndFlush(savedBook);
+    public ResponseEntity<Book> update(@PathVariable Long id, @RequestBody Book book) {
+
+        if (!bookRepository.findById(id).isPresent())
+            return new ResponseEntity("Book not found", HttpStatus.NOT_FOUND);
+        else {
+            Book savedBook = bookRepository.getById(id);
+            BeanUtils.copyProperties(book, savedBook, "id");
+            bookRepository.saveAndFlush(savedBook);
+            return new ResponseEntity(savedBook, HttpStatus.OK);
+        }
+
     }
 
 }
